pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spark-doctor"
version = "0.1.0"
description = "Local diagnostic CLI for NVIDIA DGX Spark"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Joey" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
spark-doctor = "spark_doctor.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"spark_doctor.recipes" = ["*.yaml"]
